Fix serial_putc type consistency across all Rust modules
- Update devfs.rs: i8 -> u8 for serial_putc
- Update procfs.rs: i8 -> u8 for serial_putc
- Update thread.rs: i8 -> u8 for serial_putc
- Update session.rs: i8 -> u8 for serial_putc

This eliminates all 'clashing_extern_declarations' warnings
and ensures type safety consistency with vfs.rs, diskfs.rs,
hvfs.rs, and ramfs.rs which were already fixed.

+++ src/fs/procfs/procfs.rs
@@ -2,13 +2,13 @@ use spin::Mutex;
 use core::sync::atomic::{AtomicU32, Ordering};
 
 extern "C" {
-    fn serial_putc(port: u16, c: i8);
+    fn serial_putc(port: u16, c: u8);
 }
 
 fn log(s: &str) {
     unsafe {
         for c in s.bytes() {
-            serial_putc(0x3F8, c as i8);
+            serial_putc(0x3F8, c);
         }
     }
 }
